chore: extract interfaces to make clients injectable and mockable
