+++ mailbox-sdk/src/main/java/com/zextras/mailbox/client/Client.java
@@ -1,0 +1,7 @@
+package com.zextras.mailbox.client;
+
+import com.zextras.mailbox.client.requests.Request;
+
+public interface Client<Service> {
+  <Res> Res send(Request<Service, Res> request);
+}
